requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
docxtpl==0.16.7
docx2pdf==0.1.8
